// legacy/scripts/standardize-artifacts-interface.ts
export interface BuildInfoInterface {
  id: string;
  _format: string;
  solcVersion: string;
  solcLongVersion: string;
  input: Input;
  output: Output;
}

export interface Input {
  language: string;
  sources: Record<string, any>;
  settings: Settings;
}

export interface Settings {
  evmVersion: string;
  metadata: Metadata;
  optimizer: Optimizer;
  outputSelection: OutputSelection;
}

export interface Metadata {
  useLiteralContent: boolean;
}

export interface Optimizer {
  enabled: boolean;
  runs: number;
}

export interface OutputSelection {
  "*": GeneratedType;
}

export interface GeneratedType {
  "*": string[];
  "": string[];
}

export interface Output {
  contracts: Record<string, Record<string, any>>;
  sources: Record<string, OutputSource>;
}

export interface OutputSource {
  ast: any;
  id: number;
}

export interface ArtifactInterface {
  contractName: string;
  abi: Abi[];
  metadata: string;
  bytecode: string;
  deployedBytecode: string;
  immutableReferences?: ImmutableReferences;
  generatedSources?: any[];
  deployedGeneratedSources?: any[];
  sourceMap?: string;
  deployedSourceMap?: string;
  source: string;
  sourcePath?: string;
  ast: Ast;
  compiler: Compiler;
  networks?: Networks;
  schemaVersion?: string;
  updatedAt?: string;
  devdoc?: Devdoc;
  userdoc?: Userdoc;
}

export interface Abi {
  anonymous?: boolean;
  inputs: Input[];
  name: string;
  type: string;
  outputs?: Output[];
  stateMutability?: string;
}

export interface Input {
  indexed?: boolean;
  internalType: string;
  name: string;
  type: string;
}

export interface Output {
  internalType: string;
  name: string;
  type: string;
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface ImmutableReferences {}

export interface Ast {
  absolutePath: string;
  exportedSymbols: ExportedSymbols;
  id: number;
  license: string;
  nodeType: string;
  nodes: Node[];
  src: string;
}

export interface ExportedSymbols {
  Context: number[];
  Ownable: number[];
}

export interface Node {
  id: number;
  literals?: string[];
  nodeType: string;
  src: string;
  absolutePath?: string;
  file?: string;
  scope?: number;
  sourceUnit?: number;
  symbolAliases?: any[];
  unitAlias?: string;
  abstract?: boolean;
  baseContracts?: BaseContract[];
  contractDependencies?: number[];
  contractKind?: string;
  documentation?: Documentation;
  fullyImplemented?: boolean;
  linearizedBaseContracts?: number[];
  name?: string;
  nodes?: Node2[];
}
// ...
export interface Compiler {
  name: string;
  version: string;
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface Networks {}

export interface Devdoc {
  details: string;
  kind: string;
  methods: Methods;
  version: number;
}

export interface Methods {
  constructor: Constructor;
  "owner()": Owner;
  "renounceOwnership()": RenounceOwnership;
  "transferOwnership(address)": TransferOwnershipAddress;
}

export interface Constructor {
  details: string;
}

export interface Owner {
  details: string;
}

export interface RenounceOwnership {
  details: string;
}

export interface TransferOwnershipAddress {
  details: string;
}

export interface Userdoc {
  kind: string;
  methods: Methods2;
  version: number;
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface Methods2 {}

export interface Contract {
  name: string;
  relativePath: string;
  dbg: string;
  artifact: string;
  buildInfoPath?: string;
}

export interface Dbg {
  buildInfo: string;
}